Add files via upload
Recursion in algorithms is a programming concept where a function calls itself to solve a problem. This technique is particularly useful when a problem can be broken down into smaller instances that share the same structure as the original. Recursion follows the "divide and conquer" strategy, breaking a complex problem into simpler subproblems until a base case is reached, where the solution is straightforward.